+++ src/geneticAlgorithm/selection/EliteSelection.java
@@ -10,7 +10,7 @@ public class EliteSelection implements Selectionable {
 
 	@Override
 	public List<Individual> selection(List<Individual> list, int sizeOfNewList) {
-		Collections.sort(list, (a,b) -> Integer.compare(b.getFitness(), a.getFitness()));
+		Collections.sort(list, (a,b) -> Double.compare(b.getFitness(), a.getFitness()));
 		return new ArrayList<Individual>(list.subList(0, sizeOfNewList));
 	}
 }
